Moved IImageUrlGenerator and implementation into more appropriate enamespace
