+++ src/Umbraco.Infrastructure/PropertyEditors/GridPropertyEditor.cs
@@ -5,6 +5,7 @@ using Newtonsoft.Json;
 using Umbraco.Core;
 using Umbraco.Core.IO;
 using Umbraco.Core.Logging;
+using Umbraco.Core.Media;
 using Umbraco.Core.Models;
 using Umbraco.Core.Models.Editors;
 using Umbraco.Core.PropertyEditors;
