+++ src/Umbraco.Infrastructure/PropertyEditors/RichTextPropertyEditor.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Umbraco.Core;
 using Umbraco.Core.IO;
 using Umbraco.Core.Logging;
+using Umbraco.Core.Media;
 using Umbraco.Core.Models;
 using Umbraco.Core.Models.Editors;
 using Umbraco.Core.PropertyEditors;
