+++ src/Umbraco.Tests/PublishedContent/PublishedContentTestBase.cs
@@ -16,6 +16,7 @@ using Umbraco.Web;
 using Umbraco.Web.Templates;
 using Umbraco.Web.Models;
 using Umbraco.Web.Routing;
+using Umbraco.Core.Media;
 
 namespace Umbraco.Tests.PublishedContent
 {
