+++ src/Umbraco.Tests/Web/Controllers/UsersControllerTests.cs
@@ -37,6 +37,7 @@ using IUser = Umbraco.Core.Models.Membership.IUser;
 using Umbraco.Core.Mapping;
 using Umbraco.Core.Configuration.UmbracoSettings;
 using Umbraco.Web.Routing;
+using Umbraco.Core.Media;
 
 namespace Umbraco.Tests.Web.Controllers
 {
