+++ src/Umbraco.Web/Editors/ImageUrlGeneratorController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Umbraco.Core.Models;
+using Umbraco.Core.Media;
 using Umbraco.Web.Models;
 using Umbraco.Web.Mvc;
