+++ src/Umbraco.Web/Editors/UsersController.cs
@@ -36,6 +36,7 @@ using Task = System.Threading.Tasks.Task;
 using Umbraco.Core.Mapping;
 using Umbraco.Core.Configuration.UmbracoSettings;
 using Umbraco.Web.Routing;
+using Umbraco.Core.Media;
 
 namespace Umbraco.Web.Editors
 {
